+++ docs/tasks/artifacts/sv_corpus_grad/enum_base_range/e1_dim_no_base.sv
@@ -1,0 +1,3 @@
+module m;
+  typedef enum [2:0] { A = 1, B = 2 } three_t;
+endmodule