+++ docs/tasks/artifacts/sv_corpus_grad/enum_base_range/e2_no_base_no_dim.sv
@@ -1,0 +1,3 @@
+module m;
+  typedef enum { A = 1, B = 2 } e_t;
+endmodule