+++ docs/tasks/artifacts/sv_corpus_grad/enum_base_range/e3_vector_base_dim.sv
@@ -1,0 +1,3 @@
+module m;
+  typedef enum logic [2:0] { A = 1, B = 2 } e_t;
+endmodule