+++ docs/tasks/artifacts/sv_corpus_grad/enum_base_range/e4_atom_base.sv
@@ -1,0 +1,3 @@
+module m;
+  typedef enum int { A = 1, B = 2 } e_t;
+endmodule