+++ docs/tasks/artifacts/sv_corpus_grad/enum_base_range/e5_signing_dim.sv
@@ -1,0 +1,3 @@
+module m;
+  typedef enum bit signed [2:0] { A = 1 } e_t;
+endmodule