+++ docs/tasks/artifacts/sv_corpus_grad/enum_base_range/e6_typeid_dim.sv
@@ -1,0 +1,3 @@
+module m;
+  typedef enum my_t [2:0] { A = 1 } e_t;
+endmodule